firebase refactor
