+++ src/api/square.js
@@ -1,0 +1,7 @@
+
+export default class SquareAPI {
+    constructor() {
+      
+    }
+
+}
